Rename qr_gen generator.
