+++ vision/generator/QrCode.py
