+++ vision/generator/QrCodeUsage.py
